Change VOLUME/AMOUNT from int64_t to double; fix build and tests
Core changes:
- MarketData VOLUME/AMOUNT: Column<int64_t> → Column<double>
- ParquetReader/HDF5Reader: store volume/amount as double directly,
  remove int64_t conversion (llround/static_cast) and roundVolumeAndAmount branching
- fieldIsDouble(): return true for VOLUME/AMOUNT

Build fixes:
- CompilerSettings: C++17→C++20, fix Release/Debug condition conflict
- quantcore CMakeLists: fix include path (PROJECT_SOURCE_DIR)
- FindArrow.cmake: remove double-arrow PATH_SUFFIXES bug
- HDF5Reader.cpp: add Impl definition in #else stub for unique_ptr

Test fixes:
- test_parquet_reader: add <filesystem>, int64_t→double
- test_market_data: int64_t→double, add ErrorHandling.h include,
  fix pre-existing expected value bug (2200000→2300000)
- test_hdf5_reader: update roundVolumeAndAmount default assertion
- Add testdata/hdf5_daily/20150210.h5 for HDF5 reader tests

Docs:
- README: update all VOLUME/AMOUNT type references to double
  (https://github.com/google/googletest.git)

+++ quantcore/storage/MarketData.cpp
@@ -18,8 +18,7 @@ MarketData::MarketData(std::string assetId, TimestampIndex timestamps)
     , timestamps_(std::move(timestamps))
 {
     // Initialize all field slots with empty double columns by default.
-    // Fields that need int64_t (VOLUME, AMOUNT) start empty until
-    // explicitly set.
+    // All fields are double; columns start empty until explicitly set.
 }
 
 bool MarketData::allColumnsAligned() const {
@@ -87,9 +86,9 @@ void MarketData::allocateAllFields() {
     columns_[static_cast<std::size_t>(Field::CLOSE)]  = Column<double>(rows);
     columns_[static_cast<std::size_t>(Field::VWAP)]   = Column<double>(rows);
 
-    // Quantity fields: int64_t
-    columns_[static_cast<std::size_t>(Field::VOLUME)] = Column<int64_t>(rows);
-    columns_[static_cast<std::size_t>(Field::AMOUNT)] = Column<int64_t>(rows);
+    // Quantity fields: double
+    columns_[static_cast<std::size_t>(Field::VOLUME)] = Column<double>(rows);
+    columns_[static_cast<std::size_t>(Field::AMOUNT)] = Column<double>(rows);
 }
 
 // ============================================================
